+++ CHANGELOG.markdown
@@ -1,5 +1,11 @@
+September 24, 2010
+
+* Reworked some of the sound effects. Launch weapon sound, hit sound.
+* Enemy will now hit you.
+
 September 23, 2010
 
 * MDZ: Added a `#define` flag called RQAudioPlayerPlayAudioDuringDevelopment. Lets me/us have the app stay quite during development.
 * MDZ: Added a flick threshold that will force the user to define the trajectory and velocity of the weapon before they pass the yellow line.
 * MDZ: Added some basic perspective as the weapon passes the flick threshold it will shrink. Not sure how well it works visually. Feedback welcome.
+